Feat(Commands): ping bot giveaways cmd (#50)
* Ping bot giveaways command

* Update src/commands/utility/pingbotgiveaways.js


* Update src/commands/utility/pingbotgiveaways.js


* Update src/preconditions/GiveawayHoster.js


* Update src/preconditions/GiveawayHoster.js

+++ src/listeners/command/messageCommandDenied.js
@@ -1,6 +1,8 @@
 const { Listener, Events } = require('@sapphire/framework');
+const { DurationFormatter } = require('@sapphire/time-utilities');
 const Sentry = require('@sentry/node');
 const { Message } = require('discord.js');
+const { ErrorEmbed } = require('../../library/embeds');
 
 class MessageCommandDeniedListener extends Listener {
     constructor(context, options) {
@@ -16,6 +18,19 @@ class MessageCommandDeniedListener extends Listener {
      * @param { Message } message
      */
     async run(_error, { message }) {
+        console.log(_error);
+        console.log(_error.name)
+        if(_error.precondition.name == 'Cooldown') {
+            return message.channel.send({
+                embeds: [
+                    ErrorEmbed(
+                        `You're on cooldown! Please wait **${new DurationFormatter().format(_error.context.remaining)}** before using this command again.`,
+                        message.author
+                    ),
+                ],
+            });
+        };
+
         if (message.deletable) await message.delete();
         else
             Sentry.captureMessage(
